+++ sdk/keyvault/keyvault-admin/src/backupClient.ts
@@ -101,7 +101,7 @@ export class KeyVaultBackupClient {
   }
 
   /**
-   * Starts generating a backup of an Azure Key Vault on the specified Storage Blob account using the provided SAS token to access the account.
+   * Starts generating a backup of an Azure Key Vault on the specified Storage Blob account.
    *
    * This function returns a Long Running Operation poller that allows you to wait indefinitely until the Key Vault backup is generated.
    *
@@ -128,75 +128,22 @@ export class KeyVaultBackupClient {
    * ```
    * Starts a full backup operation.
    * @param blobStorageUri - The URL of the blob storage resource, including the path to the container where the backup will end up being stored.
-   * @param sasToken - The SAS token.
+   * @param sasToken - The SAS token. If no SAS token is provided, user-assigned Managed Identity will be used to access the blob storage resource.
    * @param options - The optional parameters.
    */
   public async beginBackup(
     blobStorageUri: string,
-    sasToken: string,
-    options?: KeyVaultBeginBackupOptions
-  ): Promise<PollerLike<KeyVaultBackupOperationState, KeyVaultBackupResult>>;
-
-  /**
-   * Starts generating a backup of an Azure Key Vault on the specified Storage Blob account using a user-assigned Managed Identity to access the account.
-   *
-   * This function returns a Long Running Operation poller that allows you to wait indefinitely until the Key Vault backup is generated.
-   *
-   * Example usage:
-   * ```ts
-   * const client = new KeyVaultBackupClient(url, credentials);
-   *
-   * const blobStorageUri = "<blob-storage-uri>"; // <Blob storage URL>/<folder name>
-   * const sasToken = "<sas-token>";
-   * const poller = await client.beginBackup(blobStorageUri);
-   *
-   * // Serializing the poller
-   * //
-   * //   const serialized = poller.toString();
-   * //
-   * // A new poller can be created with:
-   * //
-   * //   await client.beginBackup(blobStorageUri, { resumeFrom: serialized });
-   * //
-   *
-   * // Waiting until it's done
-   * const backupUri = await poller.pollUntilDone();
-   * console.log(backupUri);
-   * ```
-   * Starts a full backup operation.
-   * @param blobStorageUri - The URL of the blob storage resource, including the path to the container where the backup will end up being stored.
-   * @param sasToken - The SAS token.
-   * @param options - The optional parameters.
-   */
-  public async beginBackup(
-    blobStorageUri: string,
-    options?: KeyVaultBeginBackupOptions
-  ): Promise<PollerLike<KeyVaultBackupOperationState, KeyVaultBackupResult>>;
-
-  public async beginBackup(
-    blobStorageUri: string,
-    sasTokenOrOptions: string | KeyVaultBeginBackupOptions = {},
-    options?: KeyVaultBeginBackupOptions
+    sasToken?: string,
+    options: KeyVaultBeginBackupOptions = {}
   ): Promise<PollerLike<KeyVaultBackupOperationState, KeyVaultBackupResult>> {
-    let sasToken: string | undefined;
-    let actualOptions: KeyVaultBeginBackupOptions;
-
-    if (typeof sasTokenOrOptions === "string") {
-      sasToken = sasTokenOrOptions;
-      actualOptions = options ?? {};
-    } else {
-      sasToken = undefined;
-      actualOptions = sasTokenOrOptions;
-    }
-
     const poller = new KeyVaultBackupPoller({
       blobStorageUri,
       sasToken,
       client: this.client,
       vaultUrl: this.vaultUrl,
-      intervalInMs: actualOptions.intervalInMs,
-      resumeFrom: actualOptions.resumeFrom,
-      requestOptions: actualOptions,
+      intervalInMs: options.intervalInMs,
+      resumeFrom: options.resumeFrom,
+      requestOptions: options,
     });
 
     // This will initialize the poller's operation (the generation of the backup).
@@ -234,7 +181,7 @@ export class KeyVaultBackupClient {
    * ```
    * Starts a full restore operation.
    * @param folderUri - The URL of the blob storage resource where the previous successful full backup was stored.
-   * @param sasToken - The SAS token.
+   * @param sasToken - The SAS token. If no SAS token is provided, user-assigned Managed Identity will be used to access the blob storage resource.
    * @param options - The optional parameters.
    */
   public async beginRestore(
@@ -288,7 +235,7 @@ export class KeyVaultBackupClient {
    * Creates a new role assignment.
    * @param keyName - The name of the key that wants to be restored.
    * @param folderUri - The URL of the blob storage resource, with the folder name of the blob where the previous successful full backup was stored.
-   * @param sasToken - The SAS token.
+   * @param sasToken - The SAS token. If no SAS token is provided, user-assigned Managed Identity will be used to access the blob storage resource.
    * @param options - The optional parameters.
    */
   public async beginSelectiveKeyRestore(