+++ sdk/containerservice/arm-containerservicefleet/samples-dev/fleetsListCredentialsSample.ts
@@ -18,7 +18,7 @@ dotenv.config();
  * This sample demonstrates how to Lists the user credentials of a Fleet.
  *
  * @summary Lists the user credentials of a Fleet.
- * x-ms-original-file: specification/containerservice/resource-manager/Microsoft.ContainerService/fleet/preview/2023-06-15-preview/examples/Fleets_ListCredentialsResult.json
+ * x-ms-original-file: specification/containerservice/resource-manager/Microsoft.ContainerService/fleet/preview/2023-08-15-preview/examples/Fleets_ListCredentialsResult.json
  */
 async function listsTheUserCredentialsOfAFleet() {
   const subscriptionId =