+++ sdk/monitor/monitor-opentelemetry/src/index.ts
@@ -3,6 +3,7 @@
 
 import { metrics, trace } from "@opentelemetry/api";
 import { logs } from "@opentelemetry/api-logs";
+import { NodeSDK, NodeSDKConfiguration } from "@opentelemetry/sdk-node";
 import { InternalConfig } from "./shared/config";
 import { MetricHandler } from "./metrics";
 import { TraceHandler } from "./traces/handler";
@@ -20,9 +21,7 @@ export { AzureMonitorOpenTelemetryOptions, InstrumentationOptions } from "./shar
 
 process.env["AZURE_MONITOR_DISTRO_VERSION"] = AZURE_MONITOR_OPENTELEMETRY_VERSION;
 
-let metricHandler: MetricHandler;
-let traceHandler: TraceHandler;
-let logHandler: LogHandler;
+let sdk: NodeSDK;
 
 /**
  * Initialize Azure Monitor Distro
@@ -35,23 +34,39 @@ export function useAzureMonitor(options?: AzureMonitorOpenTelemetryOptions) {
   metrics.disable();
   trace.disable();
   logs.disable();
+
   // Create internal handlers
-  metricHandler = new MetricHandler(config);
-  traceHandler = new TraceHandler(config, metricHandler);
-  logHandler = new LogHandler(config, metricHandler);
+  const metricHandler = new MetricHandler(config);
+  const traceHandler = new TraceHandler(config, metricHandler);
+  const logHandler = new LogHandler(config, metricHandler);
+
+  // Initialize OpenTelemetry SDK
+  const sdkConfig: Partial<NodeSDKConfiguration> = {
+    autoDetectResources: true,
+    logRecordProcessor: logHandler.getLogRecordProcessor(),
+    metricReader: metricHandler.getMetricReader(),
+    views: metricHandler.getViews(),
+    instrumentations: traceHandler.getInstrumentations(),
+    resource: config.resource,
+    sampler: traceHandler.getSampler(),
+    spanProcessor: traceHandler.getSpanProcessor(),
+  };
+  sdk = new NodeSDK(sdkConfig);
+  sdk.start();
+  // Add extra SpanProcessors, MetricReaders and LogRecordProcessors
+  traceHandler.start();
+  logHandler.start();
 }
 
 /**
  * Shutdown Azure Monitor Distro
  */
 export function shutdownAzureMonitor() {
-  metricHandler.shutdown();
-  traceHandler.shutdown();
-  logHandler.shutdown();
+  sdk?.shutdown();
 }
 
 function _setStatsbeatFeatures(config: InternalConfig) {
-  let instrumentationBitMap = 0;
+  let instrumentationBitMap = StatsbeatInstrumentation.NONE;
   if (config.instrumentationOptions?.azureSdk?.enabled) {
     instrumentationBitMap |= StatsbeatInstrumentation.AZURE_CORE_TRACING;
   }
@@ -68,10 +83,14 @@ function _setStatsbeatFeatures(config: InternalConfig) {
     instrumentationBitMap |= StatsbeatInstrumentation.REDIS;
   }
 
-  let featureBitMap = 0;
+  let featureBitMap = StatsbeatFeature.NONE;
   featureBitMap |= StatsbeatFeature.DISTRO;
 
   try {
+    const currentFeaturesBitMap = Number(process.env[AZURE_MONITOR_STATSBEAT_FEATURES]);
+    if (!isNaN(currentFeaturesBitMap)) {
+      featureBitMap |= currentFeaturesBitMap;
+    }
     process.env[AZURE_MONITOR_STATSBEAT_FEATURES] = JSON.stringify({
       instrumentation: instrumentationBitMap,
       feature: featureBitMap,