+++ sdk/monitor/monitor-opentelemetry/src/types.ts
@@ -1,20 +1,22 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-export const AZURE_MONITOR_OPENTELEMETRY_VERSION = "1.0.0";
+export const AZURE_MONITOR_OPENTELEMETRY_VERSION = "1.1.0";
 export const AZURE_MONITOR_STATSBEAT_FEATURES = "AZURE_MONITOR_STATSBEAT_FEATURES";
 
 export enum StatsbeatFeature {
-  DISK_RETRY = 0,
-  AAD_HANDLING = 1,
-  WEB_SNIPPET = 2,
-  DISTRO = 4,
+  NONE = 0,
+  DISK_RETRY = 1,
+  AAD_HANDLING = 2,
+  WEB_SNIPPET = 4,
+  DISTRO = 8,
 }
 
 export enum StatsbeatInstrumentation {
-  AZURE_CORE_TRACING = 0,
-  MONGODB = 1,
-  MYSQL = 2,
-  REDIS = 4,
-  POSTGRES = 8,
+  NONE = 0,
+  AZURE_CORE_TRACING = 1,
+  MONGODB = 2,
+  MYSQL = 4,
+  REDIS = 8,
+  POSTGRES = 16,
 }