+++ sdk/monitor/monitor-opentelemetry/test/internal/unit/main.test.ts
@@ -10,8 +10,19 @@ import {
   shutdownAzureMonitor,
 } from "../../../src/index";
 import { MeterProvider } from "@opentelemetry/sdk-metrics";
+import { StatsbeatFeature, StatsbeatInstrumentation } from "../../../src/types";
 
 describe("Main functions", () => {
+  let originalEnv: NodeJS.ProcessEnv;
+
+  beforeEach(() => {
+    originalEnv = process.env;
+  });
+
+  afterEach(() => {
+    process.env = originalEnv;
+  });
+
   after(() => {
     trace.disable();
     metrics.disable();
@@ -31,6 +42,12 @@ describe("Main functions", () => {
   });
 
   it("should shutdown azureMonitor", () => {
+    let config: AzureMonitorOpenTelemetryOptions = {
+      azureMonitorExporterOptions: {
+        connectionString: "InstrumentationKey=00000000-0000-0000-0000-000000000000",
+      },
+    };
+    useAzureMonitor(config);
     shutdownAzureMonitor();
     const meterProvider = metrics.getMeterProvider() as MeterProvider;
     assert.strictEqual(meterProvider["_shutdown"], true);
@@ -60,9 +77,41 @@ describe("Main functions", () => {
       },
     };
     useAzureMonitor(config);
-    assert.strictEqual(
-      process.env["AZURE_MONITOR_STATSBEAT_FEATURES"],
-      JSON.stringify({ instrumentation: 15, feature: 4 })
+    let output = JSON.parse(String(process.env["AZURE_MONITOR_STATSBEAT_FEATURES"]));
+    const features = Number(output["feature"]);
+    const instrumentations = Number(output["instrumentation"]);
+    assert.ok(!(features & StatsbeatFeature.AAD_HANDLING), "AAD_HANDLING is set");
+    assert.ok(!(features & StatsbeatFeature.DISK_RETRY), "DISK_RETRY is set");
+    assert.ok(!(features & StatsbeatFeature.WEB_SNIPPET), "WEB_SNIPPET is set");
+    assert.ok(features & StatsbeatFeature.DISTRO, "DISTRO is not set");
+    assert.ok(
+      instrumentations & StatsbeatInstrumentation.AZURE_CORE_TRACING,
+      "AZURE_CORE_TRACING not set"
     );
+    assert.ok(instrumentations & StatsbeatInstrumentation.MONGODB, "MONGODB not set");
+    assert.ok(instrumentations & StatsbeatInstrumentation.MYSQL, "MYSQL not set");
+    assert.ok(instrumentations & StatsbeatInstrumentation.POSTGRES, "POSTGRES not set");
+    assert.ok(instrumentations & StatsbeatInstrumentation.REDIS, "REDIS not set");
+  });
+
+  it("should use statsbeat features if already available", () => {
+    const env = <{ [id: string]: string }>{};
+    let current = 0;
+    current |= StatsbeatFeature.AAD_HANDLING;
+    current |= StatsbeatFeature.DISK_RETRY;
+    env.AZURE_MONITOR_STATSBEAT_FEATURES = current.toString();
+    process.env = env;
+    let config: AzureMonitorOpenTelemetryOptions = {
+      azureMonitorExporterOptions: {
+        connectionString: "InstrumentationKey=00000000-0000-0000-0000-000000000000",
+      },
+    };
+    useAzureMonitor(config);
+    let output = JSON.parse(String(process.env["AZURE_MONITOR_STATSBEAT_FEATURES"]));
+    const numberOutput = Number(output["feature"]);
+    assert.ok(numberOutput & StatsbeatFeature.AAD_HANDLING, "AAD_HANDLING not set");
+    assert.ok(numberOutput & StatsbeatFeature.DISK_RETRY, "DISK_RETRY not set");
+    assert.ok(numberOutput & StatsbeatFeature.DISTRO, "DISTRO not set");
+    assert.ok(!(numberOutput & StatsbeatFeature.WEB_SNIPPET), "WEB_SNIPPET is set");
   });
 });