+++ sdk/postgresql/arm-postgresql-flexible/samples/v8-beta/javascript/migrationsUpdateSample.js
@@ -19,14 +19,13 @@ require("dotenv").config();
  * x-ms-original-file: specification/postgresql/resource-manager/Microsoft.DBforPostgreSQL/preview/2023-03-01-preview/examples/Migrations_Cancel.json
  */
 async function cancelMigration() {
-  const subscriptionId =
-    process.env["POSTGRESQL_SUBSCRIPTION_ID"] || "00000000-0000-0000-0000-000000000000";
+  const subscriptionId = "ffffffff-ffff-ffff-ffff-ffffffffffff";
   const resourceGroupName = process.env["POSTGRESQL_RESOURCE_GROUP"] || "testrg";
   const targetDbServerName = "testtarget";
   const migrationName = "testmigration";
   const parameters = { cancel: "True" };
   const credential = new DefaultAzureCredential();
-  const client = new PostgreSQLManagementFlexibleServerClient(credential, subscriptionId);
+  const client = new PostgreSQLManagementFlexibleServerClient(credential);
   const result = await client.migrations.update(
     subscriptionId,
     resourceGroupName,
@@ -44,8 +43,7 @@ async function cancelMigration() {
  * x-ms-original-file: specification/postgresql/resource-manager/Microsoft.DBforPostgreSQL/preview/2023-03-01-preview/examples/Migrations_Update.json
  */
 async function migrationsUpdate() {
-  const subscriptionId =
-    process.env["POSTGRESQL_SUBSCRIPTION_ID"] || "00000000-0000-0000-0000-000000000000";
+  const subscriptionId = "ffffffff-ffff-ffff-ffff-ffffffffffff";
   const resourceGroupName = process.env["POSTGRESQL_RESOURCE_GROUP"] || "testrg";
   const targetDbServerName = "testtarget";
   const migrationName = "testmigration";
@@ -53,7 +51,7 @@ async function migrationsUpdate() {
     setupLogicalReplicationOnSourceDbIfNeeded: "True",
   };
   const credential = new DefaultAzureCredential();
-  const client = new PostgreSQLManagementFlexibleServerClient(credential, subscriptionId);
+  const client = new PostgreSQLManagementFlexibleServerClient(credential);
   const result = await client.migrations.update(
     subscriptionId,
     resourceGroupName,